Made it so you must be island owner to delete island

+++ src/main/java/com/peaches/epicskyblock/commands/DeleteCommand.java
@@ -20,7 +20,11 @@ public class DeleteCommand extends Command {
         Player p = (Player) sender;
         User user = User.getUser(p.getName());
         if (user.getIsland() != null) {
-            user.getIsland().delete();
+            if(user.getIsland().getOwner().equals(p)){
+                user.getIsland().delete();
+            }else{
+                sender.sendMessage("You must be the island owner to do this");
+            }
         } else {
             sender.sendMessage("You dont have an island");
         }
